requirements.txt
langchain
groq
tavily-python
requests
python-dotenv
newspaper3k
streamlit
